third_party/WebKit/LayoutTests/inspector/tracing/timeline-misc/timeline-aggregated-details-expected.txt


CallTree  Group by: None
  a: 0.000  11.125
    b: 0.000  11.125
      c: 8.250  8.250
      e: 0.000  2.875
        g: 2.875  2.875
  sin: 0.000  9.875
    f: 0.000  9.875
      l: 0.000  9.875
        a: 9.875  9.875
  x: 0.000  1.000
    y: 0.600  1.000
      z: 0.000  0.200
        w: 0.200  0.200
      w: 0.200  0.200
  recursive_a: 0.101  0.215
    recursive_b: 0.102  0.114
      recursive_a: 0.004  0.012
        recursive_b: 0.008  0.008
  f: 0.000  0.125
    l: 0.000  0.125
      a: 0.025  0.125
        Layout: 0.100  0.100

BottomUp  Group by: None
  a: 9.900  21.125
    l: 9.900  10.000
      f: 9.900  10.000
        sin: 9.875  9.875
  c: 8.250  8.250
    b: 8.250  8.250
      a: 8.250  8.250
  g: 2.875  2.875
    e: 2.875  2.875
      b: 2.875  2.875
        a: 2.875  2.875
  y: 0.600  1.000
    x: 0.600  1.000
  w: 0.400  0.400
    z: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
    y: 0.200  0.200
      x: 0.200  0.200
  recursive_b: 0.110  0.114
    recursive_a: 0.110  0.114
      recursive_b: 0.008  0.008
        recursive_a: 0.008  0.008
  recursive_a: 0.105  0.215
    recursive_b: 0.004  0.012
      recursive_a: 0.004  0.012
  Layout: 0.100  0.100
    a: 0.100  0.100
      l: 0.100  0.100
        f: 0.100  0.100

CallTree  Group by: EventName
  JavaScript: 0.101  22.340
    a: 0.000  11.125
      b: 0.000  11.125
        c: 8.250  8.250
        e: 0.000  2.875
          g: 2.875  2.875
    f: 0.000  0.125
      l: 0.000  0.125
        a: 0.025  0.125
          Layout: 0.100  0.100
    sin: 0.000  9.875
      f: 0.000  9.875
        l: 0.000  9.875
          a: 9.875  9.875
    x: 0.000  1.000
      y: 0.600  1.000
        z: 0.000  0.200
          w: 0.200  0.200
        w: 0.200  0.200
    recursive_a: 0.101  0.215
      recursive_b: 0.102  0.114
        recursive_a: 0.004  0.012
          recursive_b: 0.008  0.008

BottomUp  Group by: EventName
  JavaScript: 22.240  22.240
    c: 8.250  8.250
      b: 8.250  8.250
        a: 8.250  8.250
    a: 9.900  21.125
      l: 9.900  10.000
        f: 9.900  10.000
          sin: 9.875  9.875
    g: 2.875  2.875
      e: 2.875  2.875
        b: 2.875  2.875
          a: 2.875  2.875
    w: 0.400  0.400
      z: 0.200  0.200
        y: 0.200  0.200
          x: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
    y: 0.600  1.000
      x: 0.600  1.000
    recursive_b: 0.110  0.114
      recursive_a: 0.110  0.114
        recursive_b: 0.008  0.008
          recursive_a: 0.008  0.008
    recursive_a: 0.105  0.215
      recursive_b: 0.004  0.012
        recursive_a: 0.004  0.012
  Layout: 0.100  0.100
    Layout: 0.100  0.100
      a: 0.100  0.100
        l: 0.100  0.100
          f: 0.100  0.100

CallTree  Group by: Category
  Scripting: 0.101  22.340
    a: 0.000  11.125
      b: 0.000  11.125
        c: 8.250  8.250
        e: 0.000  2.875
          g: 2.875  2.875
    f: 0.000  0.125
      l: 0.000  0.125
        a: 0.025  0.125
          Layout: 0.100  0.100
    sin: 0.000  9.875
      f: 0.000  9.875
        l: 0.000  9.875
          a: 9.875  9.875
    x: 0.000  1.000
      y: 0.600  1.000
        z: 0.000  0.200
          w: 0.200  0.200
        w: 0.200  0.200
    recursive_a: 0.101  0.215
      recursive_b: 0.102  0.114
        recursive_a: 0.004  0.012
          recursive_b: 0.008  0.008

BottomUp  Group by: Category
  Scripting: 22.240  22.240
    c: 8.250  8.250
      b: 8.250  8.250
        a: 8.250  8.250
    a: 9.900  21.125
      l: 9.900  10.000
        f: 9.900  10.000
          sin: 9.875  9.875
    g: 2.875  2.875
      e: 2.875  2.875
        b: 2.875  2.875
          a: 2.875  2.875
    w: 0.400  0.400
      z: 0.200  0.200
        y: 0.200  0.200
          x: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
    y: 0.600  1.000
      x: 0.600  1.000
    recursive_b: 0.110  0.114
      recursive_a: 0.110  0.114
        recursive_b: 0.008  0.008
          recursive_a: 0.008  0.008
    recursive_a: 0.105  0.215
      recursive_b: 0.004  0.012
        recursive_a: 0.004  0.012
  Rendering: 0.100  0.100
    Layout: 0.100  0.100
      a: 0.100  0.100
        l: 0.100  0.100
          f: 0.100  0.100

CallTree  Group by: Domain
  unattributed: 0.000  12.250
    a: 0.000  11.125
      b: 0.000  11.125
        c: 8.250  8.250
        e: 0.000  2.875
          g: 2.875  2.875
    f: 0.000  0.125
      l: 0.000  0.125
        a: 0.025  0.125
          Layout: 0.100  0.100
    x: 0.000  1.000
      y: 0.600  1.000
        z: 0.000  0.200
          w: 0.200  0.200
        w: 0.200  0.200
  [V8 Runtime]: 0.000  9.875
    sin: 0.000  9.875
      f: 0.000  9.875
        l: 0.000  9.875
          a: 9.875  9.875
  abc.com: 0.101  0.215
    recursive_a: 0.101  0.215
      recursive_b: 0.102  0.114
        recursive_a: 0.004  0.012
          recursive_b: 0.008  0.008

BottomUp  Group by: Domain
  unattributed: 22.125  22.125
    c: 8.250  8.250
      b: 8.250  8.250
        a: 8.250  8.250
    a: 9.900  21.125
      l: 9.900  10.000
        f: 9.900  10.000
          sin: 9.875  9.875
    g: 2.875  2.875
      e: 2.875  2.875
        b: 2.875  2.875
          a: 2.875  2.875
    Layout: 0.100  0.100
      a: 0.100  0.100
        l: 0.100  0.100
          f: 0.100  0.100
    w: 0.400  0.400
      z: 0.200  0.200
        y: 0.200  0.200
          x: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
    y: 0.600  1.000
      x: 0.600  1.000
  abc.com: 0.215  0.215
    recursive_b: 0.110  0.114
      recursive_a: 0.110  0.114
        recursive_b: 0.008  0.008
          recursive_a: 0.008  0.008
    recursive_a: 0.105  0.215
      recursive_b: 0.004  0.012
        recursive_a: 0.004  0.012

CallTree  Group by: Subdomain
  unattributed: 0.000  12.250
    a: 0.000  11.125
      b: 0.000  11.125
        c: 8.250  8.250
        e: 0.000  2.875
          g: 2.875  2.875
    f: 0.000  0.125
      l: 0.000  0.125
        a: 0.025  0.125
          Layout: 0.100  0.100
    x: 0.000  1.000
      y: 0.600  1.000
        z: 0.000  0.200
          w: 0.200  0.200
        w: 0.200  0.200
  [V8 Runtime]: 0.000  9.875
    sin: 0.000  9.875
      f: 0.000  9.875
        l: 0.000  9.875
          a: 9.875  9.875
  xyz.abc.com: 0.101  0.215
    recursive_a: 0.101  0.215
      recursive_b: 0.102  0.114
        recursive_a: 0.004  0.012
          recursive_b: 0.008  0.008

BottomUp  Group by: Subdomain
  unattributed: 22.125  22.125
    c: 8.250  8.250
      b: 8.250  8.250
        a: 8.250  8.250
    a: 9.900  21.125
      l: 9.900  10.000
        f: 9.900  10.000
          sin: 9.875  9.875
    g: 2.875  2.875
      e: 2.875  2.875
        b: 2.875  2.875
          a: 2.875  2.875
    Layout: 0.100  0.100
      a: 0.100  0.100
        l: 0.100  0.100
          f: 0.100  0.100
    w: 0.400  0.400
      z: 0.200  0.200
        y: 0.200  0.200
          x: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
    y: 0.600  1.000
      x: 0.600  1.000
  xyz.abc.com: 0.215  0.215
    recursive_b: 0.110  0.114
      recursive_a: 0.110  0.114
        recursive_b: 0.008  0.008
          recursive_a: 0.008  0.008
    recursive_a: 0.105  0.215
      recursive_b: 0.004  0.012
        recursive_a: 0.004  0.012

CallTree  Group by: URL
  unattributed: 0.000  12.250
    a: 0.000  11.125
      b: 0.000  11.125
        c: 8.250  8.250
        e: 0.000  2.875
          g: 2.875  2.875
    f: 0.000  0.125
      l: 0.000  0.125
        a: 0.025  0.125
          Layout: 0.100  0.100
    x: 0.000  1.000
      y: 0.600  1.000
        z: 0.000  0.200
          w: 0.200  0.200
        w: 0.200  0.200
  native math.js: 0.000  9.875
    sin: 0.000  9.875
      f: 0.000  9.875
        l: 0.000  9.875
          a: 9.875  9.875
  http://xyz.abc.com/rec.js: 0.101  0.215
    recursive_a: 0.101  0.215
      recursive_b: 0.102  0.114
        recursive_a: 0.004  0.012
          recursive_b: 0.008  0.008

BottomUp  Group by: URL
  unattributed: 22.125  22.125
    c: 8.250  8.250
      b: 8.250  8.250
        a: 8.250  8.250
    a: 9.900  21.125
      l: 9.900  10.000
        f: 9.900  10.000
          sin: 9.875  9.875
    g: 2.875  2.875
      e: 2.875  2.875
        b: 2.875  2.875
          a: 2.875  2.875
    Layout: 0.100  0.100
      a: 0.100  0.100
        l: 0.100  0.100
          f: 0.100  0.100
    w: 0.400  0.400
      z: 0.200  0.200
        y: 0.200  0.200
          x: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
    y: 0.600  1.000
      x: 0.600  1.000
  http://xyz.abc.com/rec.js: 0.215  0.215
    recursive_b: 0.110  0.114
      recursive_a: 0.110  0.114
        recursive_b: 0.008  0.008
          recursive_a: 0.008  0.008
    recursive_a: 0.105  0.215
      recursive_b: 0.004  0.012
        recursive_a: 0.004  0.012

CallTree  Group by: Frame
  subframe-url1: 0.000  21.125
    a: 0.000  11.125
      b: 0.000  11.125
        c: 8.250  8.250
        e: 0.000  2.875
          g: 2.875  2.875
    f: 0.000  0.125
      l: 0.000  0.125
        a: 0.025  0.125
          Layout: 0.100  0.100
    sin: 0.000  9.875
      f: 0.000  9.875
        l: 0.000  9.875
          a: 9.875  9.875
  subframe-url3: 0.101  1.215
    x: 0.000  1.000
      y: 0.600  1.000
        z: 0.000  0.200
          w: 0.200  0.200
        w: 0.200  0.200
    recursive_a: 0.101  0.215
      recursive_b: 0.102  0.114
        recursive_a: 0.004  0.012
          recursive_b: 0.008  0.008

BottomUp  Group by: Frame
  subframe-url1: 21.025  21.025
    c: 8.250  8.250
      b: 8.250  8.250
        a: 8.250  8.250
    a: 9.900  21.125
      l: 9.900  10.000
        f: 9.900  10.000
          sin: 9.875  9.875
    g: 2.875  2.875
      e: 2.875  2.875
        b: 2.875  2.875
          a: 2.875  2.875
  subframe-url3: 1.215  1.215
    w: 0.400  0.400
      z: 0.200  0.200
        y: 0.200  0.200
          x: 0.200  0.200
      y: 0.200  0.200
        x: 0.200  0.200
    y: 0.600  1.000
      x: 0.600  1.000
    recursive_b: 0.110  0.114
      recursive_a: 0.110  0.114
        recursive_b: 0.008  0.008
          recursive_a: 0.008  0.008
    recursive_a: 0.105  0.215
      recursive_b: 0.004  0.012
        recursive_a: 0.004  0.012
  "subframe-name2": 0.100  0.100
    Layout: 0.100  0.100
      a: 0.100  0.100
        l: 0.100  0.100
          f: 0.100  0.100

EventLog
  Function Call: 8.875  30.000
    a: 0.000  1.000
      b: 0.000  1.000
        c: 1.000  1.000
    a: 0.000  10.125
      b: 0.000  10.125
        c: 7.250  7.250
        e: 0.000  2.875
          g: 2.875  2.875
            Timestamp: foo05: 0.000  0.000
    f: 0.000  0.125
      l: 0.000  0.125
        a: 0.025  0.125
          Layout: 0.100  0.100
          Install Timer: 0.000  0.000
    sin: 0.000  9.875
      f: 0.000  9.875
        l: 0.000  9.875
          a: 9.875  9.875
  Function Call: 8.785  10.000
    Function Call: 0.000  1.000
      x: 0.000  1.000
        y: 0.600  1.000
          z: 0.000  0.200
            w: 0.200  0.200
          w: 0.200  0.200
    recursive_a: 0.001  0.015
      recursive_b: 0.002  0.014
        recursive_a: 0.004  0.012
          recursive_b: 0.008  0.008
    recursive_a: 0.100  0.200
      recursive_b: 0.100  0.100
